暂存

+++ bos-identity/src/main/java/com/archforce/bos/identity/mapper/ResourceMapper.java
@@ -5,11 +5,11 @@ import com.archforce.bos.identity.entity.Resource;
 import java.util.List;
 
 public interface ResourceMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(Resource record);
 
-    Resource selectByPrimaryKey(Integer id);
+    Resource selectByPrimaryKey(Long id);
 
     int updateByPrimaryKey(Resource record);
